requirements.txt
# 图像切片预处理依赖包

# 核心依赖
openslide-python>=1.2.0  # 读取全扫描切片图像（NDPI、SVS等）
Pillow>=9.0.0            # 图像处理
numpy>=1.21.0            # 数值计算

# 工具库
tqdm>=4.62.0             # 进度条显示
pyyaml>=6.0              # YAML配置文件解析

# 可视化依赖
matplotlib>=3.5.0        # 用于visualize_tiles.py

# 可选依赖（用于高级功能）
# opencv-python>=4.5.0   # OpenCV图像处理（可选）
# scikit-image>=0.19.0   # 图像处理算法（可选）
